% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuestreoAleatorioSimple.R
\name{MuestreoAleatorioSimple}
\alias{MuestreoAleatorioSimple}
\title{MUESTREO ALEATORIO SIMPLE}
\usage{
MuestreoAleatorioSimple(datos, TamanoMuestra, reemplazo = FALSE)
}
\arguments{
\item{datos}{La base de datos. Clase: "data.frame" o "tibble".}

\item{TamanoMuestra}{El tamaño de la muestra que desea. Clase: "numeric".}

\item{reemplazo}{para muestreo con reemplazo. Por defecto "FALSE". Clase: "logical".}
}
\value{
Retornará una lista con 2 elementos, los índices y la muestra correspondiente.
}
\description{
La siguiente función realizará un muestreo aleatorio simple.
}
\details{
El parámetro de reemplazo, se creó pensando en la posibilidad de que, en un futuro, se requiera un
muestreo de este tipo.

Los resultados arrojados, son los índices que corresponden a las posiciones de la base de datos que fueron seleccionados de
manera aleatoria, y la muestra que es el resultado de tomar los datos ingresados, en las posiciones de los índices aleatorios.
"SubMuestreo" hace referencia al tipo de muestreo que se debe realizar dentro de cada estrato o combinación de estratos.
}
\examples{

# Tomamos una muestra aleatoria simple de la base de datos "mtcars"
MuestreoAleatorioSimple(datos = mtcars, TamanoMuestra = 10)
}
