% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuestreoEstratificado.R
\name{MuestreoEstratificado}
\alias{MuestreoEstratificado}
\title{MUESTREO ESTRATIFICADO}
\usage{
MuestreoEstratificado(datos, InfoMuestra, SubMuestreo = "simple")
}
\arguments{
\item{datos}{La base de datos a muestrear. Clase: "data.frame" o "tibble".}

\item{InfoMuestra}{La base de datos con la información del muestreo.  Clase: "data.frame" o "tibble".}

\item{SubMuestreo}{Tipo de muestreo para cada estrato. Clase: "character", solo se puede "simple" o "sistematico".}
}
\value{
Retornará una lista con 2 elementos, los índices y la muestra correspondiente.
}
\description{
La siguiente función realizara un muestreo aleatorio estratificado.
}
\details{
Es muy importante, crear la base de datos correspondiente al parámetro "InfoMuestra", puede hacerlo en Excel e
importar o directamente en R, con el fin de especificar cuáles son las variables que compondrán los estratos. Las columnas
deben tener el mismo nombre, de las variables de la base de datos a muestrear, por ejemplo:
con un solo estrato, tenemos "datos = Marco_Manzanas_2019" que tiene la columna "COD_MUNICIPIO".
Entonces "InfoMuestra" tendrá dos columnas, una "COD_MUNICIPIO" y "Cuotas". Ingresamos los estratos que queremos muestrear en
"COD_MUNICIPIO" y en la última columna del data.frame, "cuotas" la cantidad de elementos de cada estrato
(las cantidades siempre deben estar en la última columna).
Algo importante es que la función podrá tener hasta 4 estratos, el algoritmo identifica la cantidad de estratos
con la cantidad de columnas que se ingresan en "InfoMuestra".
Los resultados arrojados, son los índices que corresponden a las posiciones de la base de datos que fueron seleccionados de
manera aleatoria, y la muestra que es el resultado de tomar los datos ingresados, en las posiciones de los índices aleatorios.
"SubMuestreo" hace referencia al tipo de muestreo que se debe realizar dentro de cada estrato o combinación de estratos.
}
\examples{
# Cargue los datos "Marco_Manzanas_2019" desde SQL y llámelos como "data" cree un archivo llamado cuotas con las
# especificaciones dadas en la descripción y elija un tipo de SubMuestro.
# MuestreoEstratificado(datos = data, InfoMuestra = cuotas, SubMuestreo = "simple")
}
