% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuestreoSistematico.R
\name{MuestreoSistematico}
\alias{MuestreoSistematico}
\title{MUESTREO SISTEMÁTICO}
\usage{
MuestreoSistematico(datos, TamanoMuestra)
}
\arguments{
\item{datos}{La base de datos. clase: "data.frame" o "tibble".}

\item{TamanoMuestra}{El tamaño de la muestra que desea. clase: "numeric" o "integer".}
}
\value{
Retornará una lista con 2 elementos, los índices y la muestra correspondiente.
}
\description{
La siguiente función realizara un muestreo aleatorio sistemático.
}
\details{
En ocasiones se requiere ordenar los datos por alguna/s variable/s, para realizar una muestra que sea más representativa
a algunas características de la población como el estrato socioeconómico, es necesario que se ordene el data frame antes
de ingresarlo a la función. Una opción es utilizar dplyr::arrange().

Los resultados arrojados, son los índices que corresponden a las posiciones de la base de datos que fueron seleccionados de
manera aleatoria, y la muestra que es el resultado de tomar los datos ingresados, en las posiciones de los índices aleatorios.
"SubMuestreo" hace referencia al tipo de muestreo que se debe realizar dentro de cada estrato o combinación de estratos.
}
\examples{
# Tomamos una muestra aleatoria sistemática de la base de datos "mtcars"
MuestreoSistematico(datos = mtcars, 10)
}
